pyproject.toml
[tool.poetry]
name = "minimal-serverless-llm-deploment"
version = "0.1.0"
description = "Serverless LLMs as microservices with Replicate."
authors = ["tleers <mail@timleers.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.0.198"
fastapi = "^0.97.0"
uvicorn = "^0.22.0"
python-dotenv = "^1.0.0"
replicate = "^0.8.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-mock = "^3.10.0"
black = "^23.3.0"
isort = "^5.12.0"
autoflake = "^2.1.1"
mypy = "^1.3.0"
pre-commit = "^3.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
